+++ app/models/offer_model.py
@@ -9,7 +9,7 @@ from models.base_models import BaseSQLModel, DateCreatedChangedBase
 
 if TYPE_CHECKING:
     from models import ServiceNameModel
-    from entities.users.model_user import EmployeeModel
+    from entities.users.models_user import EmployeeModel
 
 
 class OfferLinkInSchema(BaseSQLModel):
