+++ app/routers/choices.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from entities.users.model_user import ClientModel, EmployeeModel
+from entities.users.models_user import ClientModel, EmployeeModel
 
 
 class OrderFilter(str, Enum):
