+++ app/services/employee_service.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from entities.users.model_user import EmployeeModel
+from entities.users.models_user import EmployeeModel
 from services.base_service import AbstractService
